# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Canvica Cleaning Services CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, prospect, vip]
        - name: propertyType
          in: query
          schema:
            type: string
            enum: [residential, commercial]
        - name: search
          in: query
          schema:
            type: string
        - name: recurring
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID with full profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /customers/{id}/timeline:
    get:
      operationId: getCustomerTimeline
      tags: [customers]
      summary: Get customer activity timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Activity events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
  /customers/{id}/ai-summary:
    get:
      operationId: getCustomerAiSummary
      tags: [customers]
      summary: Get AI-generated customer summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: AI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSummary"

  # ─── JOBS ───────────────────────────────────────────────────────────────────
...
  /communications:
    get:
      operationId: listCommunications
      tags: [communications]
      summary: List communications
      parameters:
        - name: customerId
          in: query
          schema:
            type: number
        - name: type
          in: query
          schema:
            type: string
            enum: [email, phone, sms, note, meeting]
      responses:
        "200":
          description: Communications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Communication"
    post:
      operationId: createCommunication
      tags: [communications]
      summary: Log a communication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunicationInput"
      responses:
        "201":
          description: Created communication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Communication"
  /communications/{id}:
    delete:
      operationId: deleteCommunication
      tags: [communications]
      summary: Delete a communication log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Customer:
      type: object
      required: [id, name, email, phone, status, serviceType, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        gpsLat:
          type: ["number", "null"]
        gpsLng:
          type: ["number", "null"]
        serviceType:
          type: string
          enum: [residential, commercial]
        propertyType:
          type: ["string", "null"]
        bedrooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["integer", "null"]
        squareFootage:
          type: ["integer", "null"]
        entryInstructions:
          type: ["string", "null"]
        alarmCode:
          type: ["string", "null"]
        gateCode:
          type: ["string", "null"]
        pets:
          type: ["string", "null"]
        preferredProducts:
          type: ["string", "null"]
        cleaningPreferences:
          type: ["string", "null"]
        specialInstructions:
          type: ["string", "null"]
        recurringSchedule:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, prospect, vip]
        referralSource:
          type: ["string", "null"]
        lifetimeValue:
          type: number
        isRecurring:
          type: boolean
        satisfactionScore:
          type: ["number", "null"]
        reviewScore:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CustomerDetail:
      type: object
      required: [id, name, email, phone, status, serviceType, createdAt, recentJobs, recentCommunications, openInvoices]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        gpsLat:
          type: ["number", "null"]
        gpsLng:
          type: ["number", "null"]
        serviceType:
          type: string
        propertyType:
          type: ["string", "null"]
        bedrooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["integer", "null"]
        squareFootage:
          type: ["integer", "null"]
        entryInstructions:
          type: ["string", "null"]
        alarmCode:
          type: ["string", "null"]
        gateCode:
          type: ["string", "null"]
        pets:
          type: ["string", "null"]
        preferredProducts:
          type: ["string", "null"]
        cleaningPreferences:
          type: ["string", "null"]
        specialInstructions:
          type: ["string", "null"]
        recurringSchedule:
          type: ["string", "null"]
        status:
          type: string
        referralSource:
          type: ["string", "null"]
        lifetimeValue:
          type: number
        isRecurring:
          type: boolean
        satisfactionScore:
          type: ["number", "null"]
        reviewScore:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        recentJobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        recentCommunications:
          type: array
          items:
            $ref: "#/components/schemas/Communication"
        openInvoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        totalJobsCount:
          type: number
        aiSummary:
          type: ["string", "null"]
    CustomerInput:
      type: object
      required: [name, email, phone, serviceType]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        gpsLat:
          type: number
        gpsLng:
          type: number
        serviceType:
          type: string
          enum: [residential, commercial]
        propertyType:
          type: string
        bedrooms:
          type: number
        bathrooms:
          type: number
        squareFootage:
          type: number
        entryInstructions:
          type: string
        alarmCode:
          type: string
        gateCode:
          type: string
        pets:
          type: string
        preferredProducts:
          type: string
        cleaningPreferences:
          type: string
        specialInstructions:
          type: string
        recurringSchedule:
          type: string
        status:
          type: string
          enum: [active, inactive, prospect, vip]
        referralSource:
          type: string
        isRecurring:
          type: boolean
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        gpsLat:
          type: number
        gpsLng:
          type: number
        serviceType:
          type: string
          enum: [residential, commercial]
        propertyType:
          type: string
        bedrooms:
          type: number
        bathrooms:
          type: number
        squareFootage:
          type: number
        entryInstructions:
          type: string
        alarmCode:
          type: string
        gateCode:
          type: string
        pets:
          type: string
        preferredProducts:
          type: string
        cleaningPreferences:
          type: string
        specialInstructions:
          type: string
        recurringSchedule:
          type: string
        status:
          type: string
          enum: [active, inactive, prospect, vip]
        referralSource:
          type: string
        isRecurring:
          type: boolean
        satisfactionScore:
          type: number
        reviewScore:
          type: number
        notes:
          type: string
    TimelineEvent:
      type: object
      required: [id, type, title, description, date]
      properties:
        id:
          type: number
        type:
          type: string
          enum: [job, quote, invoice, communication, note]
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        amount:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
    AiSummary:
      type: object
      required: [summary, keyPoints, recommendations]
      properties:
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        lastUpdated:
          type: string
          format: date-time

    # ─── JOB SCHEMAS ──────────────────────────────────────────────────────────
    Job:
      type: object
      required: [id, customerId, customerName, serviceType, cleaningType, status, scheduledDate, scheduledTime, price, createdAt]
      properties:
        id:
          type: number
        customerId:
          type: number
        customerName:
          type: string
        customerAddress:
          type: ["string", "null"]
        serviceType:
          type: string
          enum: [residential, commercial]
        cleaningType:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled, late]
        scheduledDate:
          type: string
          format: date
        scheduledTime:
          type: string
        durationHours:
          type: ["number", "null"]
        price:
          type: number
        assignedEmployeeIds:
          type: array
          items:
            type: number
        assignedEmployeeNames:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        completionNotes:
          type: ["string", "null"]
        customerRating:
          type: ["number", "null"]
        isRecurring:
          type: boolean
        recurrencePattern:
          type: ["string", "null"]
        invoiceId:
          type: ["integer", "null"]
        quoteId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    QuoteLineItem:
      type: object
      required: [description, quantity, unitPrice, total]
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
...
    Invoice:
      type: object
      required: [id, customerId, customerName, status, lineItems, subtotal, tax, total, dueDate, createdAt]
      properties:
        id:
          type: number
        invoiceNumber:
          type: string
        customerId:
          type: number
        customerName:
          type: string
        customerEmail:
          type: ["string", "null"]
        jobId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        subtotal:
          type: number
        taxRate:
          type: number
        tax:
          type: number
        total:
          type: number
        paidAmount:
          type: number
        paymentMethod:
          type: ["string", "null"]
        paymentDate:
          type: ["string", "null"]
          format: date
        dueDate:
          type: string
          format: date
        sentAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Communication:
      type: object
      required: [id, customerId, type, subject, content, direction, createdAt]
      properties:
        id:
          type: number
        customerId:
          type: number
        customerName:
          type: ["string", "null"]
        type:
          type: string
          enum: [email, phone, sms, note, meeting]
        subject:
          type: string
        content:
          type: string
        direction:
          type: string
          enum: [inbound, outbound, internal]
        staffName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CommunicationInput:
      type: object
      required: [customerId, type, subject, content, direction]
      properties:
        customerId:
          type: number
        type:
          type: string
          enum: [email, phone, sms, note, meeting]
        subject:
          type: string
        content:
          type: string
        direction:
          type: string
          enum: [inbound, outbound, internal]
        staffName:
          type: string